DRF-02/drf/requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
dj-database-url==3.0.1
Django==5.0.14
django-filter==25.1
djangorestframework==3.16.0
EditorConfig==0.17.1
gunicorn==23.0.0
idna==3.10
jsbeautifier==1.15.4
json5==0.12.0
Markdown==3.8.2
packaging==25.0
pathspec==0.12.1
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
